requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
python-dotenv
pydantic
pandas
openpyxl
pydantic[email]
